# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: Marvol Facility Cleaning Management API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /tasks:
    get:
      operationId: listTasks
      tags: [tasks]
      summary: List all tasks with optional filters
      parameters:
        - name: areaId
          in: query
          schema:
            type: integer
        - name: date
          in: query
          schema:
            type: string
        - name: assignedToId
          in: query
          schema:
            type: integer
      responses:
        "200":
          description: List of tasks
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Task"
  /tasks/special:
    get:
      operationId: listSpecialTasks
      tags: [tasks]
      summary: List special-request tasks for a date and optional area
      parameters:
        - name: date
          in: query
          schema:
            type: string
        - name: areaId
          in: query
          schema:
            type: integer
      responses:
        "200":
          description: List of special-request tasks
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/SpecialTask"
    post:
      operationId: createSpecialTask
      tags: [tasks]
      summary: Create a special-request task
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateSpecialTaskRequest"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SpecialTask"
        "400":
          description: Validation error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorEnvelope"
  /tasks/{id}/complete:
    post:
      operationId: completeTask
      tags: [tasks]
      summary: Mark a task as complete
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CompleteTaskRequest"
      responses:
        "200":
          description: Task completed
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Task"
  /tasks/{id}/uncomplete:
    post:
      operationId: uncompleteTask
      tags: [tasks]
      summary: Mark a task as incomplete
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Task uncompleted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Task"
  /tasks/complete-all:
    post:
      operationId: completeAllTasks
      tags: [tasks]
      summary: Complete all tasks for an area on a date
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CompleteAllTasksRequest"
      responses:
        "200":
          description: All tasks completed
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/BatchCompleteResponse"
...
  /dashboard:
    get:
      operationId: getDashboard
      tags: [tasks]
      summary: Get dashboard statistics
      parameters:
        - name: date
          in: query
          schema:
            type: string
      responses:
        "200":
          description: Dashboard stats
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DashboardStats"
...
components:
  securitySchemes:
    InternalCronSecret:
      type: apiKey
      in: header
      name: X-Internal-Cron-Secret
    SendGridInboundSecret:
      type: apiKey
      in: header
      name: X-SendGrid-Inbound-Secret
  schemas:
...
    Task:
      type: object
      properties:
        id:
          type: integer
        areaId:
          type: integer
        taskDate:
          type: string
        taskName:
          type: string
        taskOrder:
          type: integer
        completed:
          type: boolean
        completedAt:
          type: string
          nullable: true
        completedById:
          type: integer
          nullable: true
        completedByName:
          type: string
          nullable: true
        completedByActive:
          type: boolean
          nullable: true
          description: Whether the staff member who completed the task is still active. Null when no completer is recorded.
        assignedToId:
          type: integer
          nullable: true
        assignedToName:
          type: string
          nullable: true
        assignedToActive:
          type: boolean
          nullable: true
          description: Whether the assigned staff member is still active. Null when no one is assigned.
        inspectorWorkflowTaskId:
          type: integer
          nullable: true
          description: Inspector workflow link when this task originated from an inspector reply.
        isSpecial:
          type: boolean
        notes:
          type: string
          nullable: true
      required:
        - id
        - areaId
        - taskDate
        - taskName
        - taskOrder
        - completed
        - isSpecial
        - inspectorWorkflowTaskId
    SpecialTask:
      type: object
      properties:
        id:
          type: integer
        areaId:
          type: integer
        areaName:
          type: string
        taskDate:
          type: string
        taskName:
          type: string
        taskOrder:
          type: integer
        completed:
          type: boolean
        completedAt:
          type: string
          nullable: true
        completedById:
          type: integer
          nullable: true
        completedByName:
          type: string
          nullable: true
        completedByActive:
          type: boolean
          nullable: true
          description: Whether the staff member who completed the task is still active. Null when no completer is recorded.
        createdById:
          type: integer
          nullable: true
        createdByName:
          type: string
          nullable: true
        createdByActive:
          type: boolean
          nullable: true
          description: Whether the staff member who created the special request is still active. Null when no creator is recorded.
        createdAt:
          type: string
        notes:
          type: string
          nullable: true
      required:
        - id
        - areaId
        - areaName
        - taskDate
        - taskName
        - taskOrder
        - completed
        - createdAt
    CreateSpecialTaskRequest:
      type: object
      properties:
        areaId:
          type: integer
        date:
          type: string
        notes:
          type: string
          minLength: 1
        createdById:
          type: integer
          nullable: true
      required:
        - areaId
        - date
        - notes
    CompleteTaskRequest:
      type: object
      properties:
        completedById:
          type: integer
      required:
        - completedById
    CompleteAllTasksRequest:
      type: object
      properties:
        areaId:
          type: integer
        date:
          type: string
        completedById:
          type: integer
      required:
        - areaId
        - date
        - completedById
    BatchCompleteResponse:
      type: object
      properties:
        updated:
          type: integer
      required:
        - updated
...
    ErrorEnvelope:
      type: object
      required: [error]
      properties:
        error:
          type: string
    DashboardStats:
      type: object
      properties:
        date:
          type: string
        totalTasks:
          type: integer
        completedTasks:
          type: integer
        totalAreas:
          type: integer
        completedAreas:
          type: integer
        openIssues:
          type: integer
        areaProgress:
          type: array
          items:
            $ref: "#/components/schemas/AreaProgress"
      required:
        - date
        - totalTasks
        - completedTasks
        - totalAreas
        - completedAreas
        - openIssues
        - areaProgress
    AreaProgress:
      type: object
      properties:
        areaId:
          type: integer
        areaName:
          type: string
        terminal:
          type: string
        totalTasks:
          type: integer
        completedTasks:
          type: integer
        percentage:
          type: number
        assignedStaff:
          type: array
          items:
            type: string
      required:
        - areaId
        - areaName
        - terminal
        - totalTasks
        - completedTasks
        - percentage
        - assignedStaff
...